add file and fix

+++ TileMap.cpp
@@ -4,12 +4,12 @@
 void CTileMap::Render() {
 	CSprites* sprites = CSprites::GetInstance();
 
-	sprites->Get(id)->Draw(indexY * TILE_HEIGHT, indexX * TILE_WIDTH);
+	sprites->Get(id)->Draw((float)indexY * TILE_HEIGHT, (float)indexX * TILE_WIDTH);
 }
 
 
 
-CTileMap::CTileMap(int id, float x, float y)
+CTileMap::CTileMap(int id, int x, int y)
 {
 	this->id = id;
 	this->indexX = x;
